Move OTP generation methods into separate class.
They don't belong in Google_Authenticator_Per_User_Prompt_Acceptance_Tests because they're not tests, they're helpers for tests.

They really belong in WebHelper, but I don't think Codeception provides a way for a helper method to return a value to the suite, so putting them in their own class is the best choice.


git-svn-id: https://plugins.svn.wordpress.org/google-authenticator-per-user-prompt/trunk@912371 b8457f37-d9ea-0310-8a92-e5e31aec5664