Remove note about potential nonce expiration bug.
There is no bug; this is the expected behavior. The check for an expired nonce runs after the check for a valid OTP; so, if the OTP has expired, then the user is redirected to the login page and shown the OTP error, and the nonce is never checked. The OTP error short-circuits the nonce-expiration error.


git-svn-id: https://plugins.svn.wordpress.org/google-authenticator-per-user-prompt/trunk@912481 b8457f37-d9ea-0310-8a92-e5e31aec5664